refactor: Move tests to different modules with support for internal stage
